Rename icon files to PascalCase
