+++ src/components/icons/Coin.tsx
