+++ src/components/icons/PublicPriceIcon.tsx
