+++ src/components/icons/StationIconLegacy.tsx
