+++ src/components/icons/index.ts
@@ -1,10 +1,10 @@
 export { CardIcon } from './CardIcon'
-export { CheckIcon } from './checkIcon'
+export { CheckIcon } from './CheckIcon'
 export { ChevronDownIcon } from './ChevronDownIcon'
 export { ChevronLeftIcon } from './ChevronLeftIcon'
 export { ChevronRightIcon } from './ChevronRightIcon'
 export { ClockIcon } from './ClockIcon'
-export { default as CoinSolid } from './coin'
+export { default as CoinSolid } from './Coin'
 export { DeleteIcon } from './DeleteIcon'
 export { DollarIcon } from './DollarIcon'
 export { EditIcon } from './EditIcon'
@@ -14,12 +14,12 @@ export { ImageUploadFormIcon } from './ImageUploadFormIcon'
 export { LayersIcon } from './LayersIcon'
 export { OneChargeHeader } from './OneChargeHeader'
 export { OneChargeLogoSecondary } from './OneChargeLogoSecondary'
-export { PackageIcon } from './package-icon'
-export { PublicPriceIcon } from './public-price-icon'
+export { PackageIcon } from './PackageIcon'
+export { PublicPriceIcon } from './PublicPriceIcon'
 export { RevenueCardIcon } from './RevenueCardIcon'
 export { RevenueHeaderIcon } from './RevenueHeaderIcon'
 export { SearchIcon } from './SearchIcon'
-export { StationIcon } from './station-icon'
+export { StationIcon } from './StationIconLegacy'
 export { default as TopUpCardIcon } from './TopUpCardIcon'
 export { UploadProfile } from './UploadProfile'
 export { UsersIcon } from './UsersIcon'
